Periodically clean outdated session cache files
When a user logs out, the corresponding web session cache file is
deleted. When the web session expires, however, the file remains in the
cache directory. In a busy Gerrit server using this plugin, the size of
the cache directory can reach hundreds of megabytes within a few months.

Remove outdated session cache files in a periodic way. Cache files are
considered as outdated when its expiration date has passed.

Change-Id: Id195f2b04ef4b323ce5c5a1c024e62ebeea39720