Periodically clean outdated session cache files
When a user logs out, the corresponding web session cache file is
deleted. When the web session expires, however, the file remains in the
cache directory. In a busy Gerrit server using this plugin, the size of
the cache directory can reach hundreds of megabytes in a few months.

Remove outdated session cache files in a periodic way. Cache files are
considered as outdated after twice the maxAge time stated in the Gerrit
config subsection [cache "web_sessions"]. The calculation is done based
on files' lastAccessTime attribute.

Change-Id: I903c502f8215b95a1b275005165604f132fe07aa