Apply changes to file-based websession cache from HA-plugin
The HA-plugin contains a modified version of the websession-flatfile
plugin. The version used in the HA-plugin was updated since it was
forked, but those changes were not applied to the websession-flatfile
plugin itself.

This change applies some minor changes made to the corresponding code
in the HA-plugin to the websession-flatfile plugin. This includes:

- Variable renaming
  - `dir` -> `websessionsDir`
  - `flatFileWebSessionCache` -> `cache`
  - `cleanupInterval` -> `cleanupIntervalMillis`
- Removing `final` from parameters in `FlatFileWebSession`
- Using `SECONDS.toMillis(1)` inline as initial delay for cleanup tasks
  instead of a class field (`INITIAL_DELAY_MS = 1000`)
- Using `path.toFile().exists()` instead of `Files.exists(path)`.

While minor changes, they bring the code closer to what is used in the
HA-plugin, making it easier to port future changes.

Change-Id: I38da3e9fc7dc2dcbada82b63d51acd595f7b93a8