Apply changes to file-based websession cache from HA-plugin
The HA-plugin contains a modified version of the websession-flatfile
plugin. The version used in the HA-plugin was updated since it was
forked, but those changes were not applied to the websession-flatfile
plugin itself.

This change applies changes made to the corresponding code in the
HA-plugin and that were not exclusively needed in the HA-plugin to
the websession-flatfile plugin. This includes:

- Switching from slf4j.Logger to FluentLogger
- Addition of additional unit tests
- Change implementation how tests manipulate the system time
- Refactoring the FlatFileWebSessionCacheCleaner
- Some code style adaptations

And excludes:

- Renaming of files/classes
- Configuration handling

This change ensures that this plugin does not converge to far from the
fork in the HA-plugin and stays compatible with newer Gerrit-versions.

Change-Id: I38da3e9fc7dc2dcbada82b63d51acd595f7b93a8