Document that cache cleaning intervals < 1h are not supported
The documentation stated that time intervals in a second or minute range
could be used to define the interval, in which the web session cache
would be cleaned. In reality all time intervals were rounded to hours.
In the case of `5m` the interval was rounded to `0`, which caused an
error, when loading the plugin.

This change corrects the documentation. A different approach would be to
allow smaller time intervals, but since cleaning the cache in too small
intervals would in reality not make sense, this was not done.

Change-Id: Id5b76d8a89e6e2d9b777a6a5b7206540f98ca74b