Adapt system time manipulation to method used in HA-plugin
The HA-plugin uses a different way of accessing and manipulating the
system time, which is needed for cache cleanup scheduling.

This change adapts the system time handling of the websession-flatfile
plugin to the one of the HA-plugin to increase consistency and thereby
will make it easier to exchange future changes between the plugins.

Change-Id: Ic3e24cb476f8dd99337fe73616012952d3945ba8