Refactor test key assignment in cache tests
Keys used for tests were either assigned to fields in the constructor or
in test-methods. In case of `newkey`, this led to some code duplication.

Following the example in the HA-plugin, the assignment of the test keys
to variables was moved to the declaration of the fields and the field
names were changed to make the code more readable.

Change-Id: I16a7e2d3cf2a378bac422a0170f10aa7244a027f