Update cache cleaner to HA-plugin version and add tests
The HA-plugin implemented tests for the cache cleaner. These tests
however relied on a refactored version of the cache cleaner that allowed
to use a mocked cleanup task.

To be able to make use of the tests and to be closer to the HA-plugin
version of the cache cleaner, this change implements the refactoring and
adds the tests.

Change-Id: I72f65210d34049ae72baa094fa24bbe62f7db0f6