Update patch set 2
Patch Set 2:

It seems this is a non backwards compatible change. It would be good to have at least two branches in this project, one for stuff that works with older Gerrit, and one for managing tip. If we apply this to master and expect master to apply to the tip Gerrit, then I suspect we need to fork a stable branch before this and merge many of the other changes after this to the stable branch.

Patch-set: 2
CC: Gerrit User 1003883 <1003883@173816e5-2b9a-37c3-8a2e-48639d4f1153>