Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Can you rebase this on the head of stable-2.15?

Actually, as I commented before, my intention was to move the changes I did on stable-2.14 to stable-2.15. Since stable-2.14 and stable-2.15 were created on the same commit the merge was just a fast-forward and no merge-commit was created. This change fixes an incompatibility in the newly added tests, which were introduced with stable-2.15. Thus, the parent has to be the currently last commit of stable-2.14.
If I get it correctly, submitting this should cause Gerrit to do a fast-forward on stable-2.15 and then to add this commit. Am I wrong in that?
Is it possible to move the stable-2.15 branch to the same commit as stable-2.14 in a different way? I guess that would involve circumventing the review process, so I did not want to do that.

Patch-set: 1
Reviewer: Gerrit User 1065256 <1065256@173816e5-2b9a-37c3-8a2e-48639d4f1153>