Clean up stale websessions based on file stats
Before this change the websession cleanup routine would load in memory
the entire list of existing websession files, just to then immediately
filter them by unboxing each one and check the value of `expireAt`.

Whilst this is doable for a small to moderate number of websessions,
does not really scale when millions of websessions need to be traversed,
especially when this needs to happen over slow file systems such as NFS.

Clean up websessions based on the last modified timestamp instead and
avoid to load into memory websessions that cannot be stale by just
looking at their file stats.

Bug: Issue 15562
Change-Id: Id345d2e89d4194d8c2e5d7982bb5fa23acf71f45